스크랩 조회(멀티건) 카운트 단위 테스트

+++ src/main/java/egovframework/com/cop/scp/service/impl/EgovArticleScrapDAO.java
@@ -17,7 +17,7 @@ public class EgovArticleScrapDAO extends EgovComAbstractDAO{
 	}
 
 	public int selectArticleScrapListCnt(ScrapVO scrapVO) {
-		return (Integer)selectOne("ArticleScrap.selectArticleScrapListCnt", scrapVO);
+		return selectOne("ArticleScrap.selectArticleScrapListCnt", scrapVO);
 	}
 
 	public void insertArticleScrap(Scrap scrap) {
